fix:promote subclass RESULT to common class
