+++ app/src/main/java/com/app/nao/photorecon/Result.java
@@ -1,0 +1,28 @@
+package com.app.nao.photorecon;
+
+import android.graphics.Rect;
+
+public class Result {
+
+
+    int classIndex;
+    Float score;
+    Rect rect;
+
+    public Result(int cls, Float output, Rect rect) {
+        this.classIndex = cls;
+        this.score = output;
+        this.rect = rect;
+    }
+    public int getClassIndex() {
+        return classIndex;
+    }
+
+    public Float getScore() {
+        return score;
+    }
+
+    public Rect getRect() {
+        return rect;
+    }
+};
